prevent use of I2S buses for NeoPixelBus
as suggested by blazoncek in https://github.com/blazoncek/WLED/issues/33

"I2S is used after all RMTs are used. That effectively means 9 or 10 LED strips (buses). As it is unlikely a user may have >8 buses (not impossible, though) it may be enough just inform a user of such conflict as a first step in resolving this."